requirements.txt
pandas
transformers==2.8.0
dask
googletrans
joblib
mlflow
numpy
tensorflow
termcolor
memoized_property
xgboost
nlp
category_encoders
psutil
protobuf
scikit_learn
